Alifuat

+++ src/main/java/com/realestate/controller/ReportController.java
@@ -27,6 +27,11 @@ public class ReportController {
     private final ReportService reportService;
 
 
+    
+        // it will get tour requests for ADMIN,MANAGER ----G05 //
+    
+
+
 
     //  It will get some statistics....   G01.................\\
     @PreAuthorize("hasAnyAuthority('ADMIN','MANAGER')")
@@ -37,6 +42,7 @@ public class ReportController {
 
 
     // it will get tour requests for ADMIN,MANAGER ----G05 //
+
     @PreAuthorize("hasAnyAuthority('ADMIN,MANAGER')")
     @GetMapping("/tour-requests")
     public ResponseMessage<List<TourRequestResponse>> getTourRequestsReport(
@@ -47,16 +53,15 @@ public class ReportController {
         return reportService.getTourRequestsReport(date1, date2, status);
     }
 
-   
 
-    /*
+
     @GetMapping("/most-popular-properties")
     @PreAuthorize("hasAnyAuthority('ADMIN','MANAGER')")
-    public ResponseEntity<List<Advert>> getMostPopularProperties(@RequestParam Integer amount) {
+    public ResponseEntity<List<Advert>> getMostPopularProperties(@RequestParam Long amount) {
         List<Advert> mostPopularProperties = reportService.getMostPopularProperties(amount);
         return new ResponseEntity<>(mostPopularProperties, HttpStatus.OK);
     }
-    */
+    
 
 
     } 
@@ -67,9 +72,4 @@ public class ReportController {
     
     
     
-    
-    
-    
-    
-    
-
+
